requirements.txt
Flask==3.0.0
mcrcon==0.7.0
python-dotenv==1.0.0
